Introduce fine-grained placeholder effects
Got rid of the `Concurrency` effect, which was too broad, and would have
been hard to interpret as an actual effect. This commit replaces it with
several different placeholder effects, mostly having to do with making
HTTP requests.

Also, the `Galley0` type synomym now refers to `Galley` without any
effects, so `runGalley` and related functions now take a `Galley
GalleyEffects` instead.

`Galley0` has a `MonadUnliftIO` instance, so it can be used as a
temporary crutch to get access to async primitives. Those need to be run
in `Galley0`, and finally lifted to a general `Galley r` monad.
Eventually, the `Galley0` actions will simply be replaced by effect
actions, and the code actually using `MonadUnliftIO` will be relegated
to interpreters.